pyproject.toml
[tool.poetry]
name = "phaethon"
version = "0.1.0"
description = "Coupling magma oceans and their atmospheres"
authors = ["Fabian L. Seidler <fabian.seidler@eaps.ethz.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
astropy = "*"
helios = { git = "git@github.com:ExPlanetology/HELIOS.git" }
numpy = "^1.26.4"
pandas = "^2.2.1"
scipy = "^1.12.0"
molmass = "^2023.8.30"
uncertainties = "^3.1.7"
matplotlib = "^3.9.2"
matplotlib-label-lines = "^0.7.0"
pyfastchem = "^3.1.1"
h5py = "^3.12.1"
tqdm = "^4.66.5"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
